change company name

+++ src/pages/api/company/[cid].ts
@@ -7,21 +7,32 @@ import { ObjectId } from 'mongodb'
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   const { cid } = req.query
   const token = req.headers.authorization
-
   try {
     const userId = jwt.verify(token!, process.env.JWT_SECRET!)
     const company = await Company.findOne({
       userId: new ObjectId(userId.id),
       _id: new ObjectId(cid as string)
     })
-
     if (!company) {
       res.status(404).send('Not found')
       return
     }
-    await Company.deleteOne({ _id: new ObjectId(cid as string) })
+    if (req.method === 'DELETE') await Company.deleteOne({ _id: new ObjectId(cid as string) })
+    else if (req.method === 'PATCH') {
+      if (!req.body.name) {
+        res.status(400).send('Name must be provided')
+        return
+      }
+      await Company.updateOne(
+        { _id: new ObjectId(cid as string) },
+        { $set: { name: req.body.name } }
+      )
+      res.status(200).json({ name: req.body.name })
+      return
+    }
   } catch (e) {
     res.status(400).send(e.message)
+    return
   }
-  res.status(200).json('Deleted')
+  res.status(200).send('Deleted')
 }
